--story=875568911 软件源环境升级到最新代码
--story=875568911 针对软件源的鉴权根据配置判断是否进行加载

+++ src/backend/maven/biz-maven/src/main/kotlin/com/tencent/bkrepo/maven/service/MavenMetadataService.kt
@@ -95,16 +95,18 @@ class MavenMetadataService(
 
     fun delete(mavenArtifactInfo: ArtifactInfo, node: NodeDetail? = null, mavenGavc: MavenGAVC? = null) {
         node?.let {
-            val (criteria, _) = nodeCriteria(
+            val (criteria, mavenVersion) = nodeCriteria(
                 projectId = node.projectId,
                 repoName = node.repoName,
                 metadata = node.metadata,
                 fullPath = node.fullPath
             )
-            criteria?.let {
-                val query = Query(criteria)
-                mavenMetadataDao.remove(query)
+            if (criteria == null) return
+            mavenVersion?.timestamp?.let {
+                criteria.and(TMavenMetadataRecord::timestamp.name).`is`(mavenVersion.timestamp)
             }
+            val query = Query(criteria)
+            mavenMetadataDao.remove(query)
         }
         mavenGavc?.let {
             val groupId = mavenGavc.groupId
